requirements-test.txt
# Test dependencies for DVSwitch
pytest>=7.0.0
pytest-xdist>=3.0.0
pytest-timeout>=2.0.0
pytest-html>=3.0.0
docker>=6.0.0
websocket-client>=1.8.0
